test(engine): update and create tests
Move tests for BATCH SIZE EXCEEDING into MultiPartitionRejectionTest
class
