+++ engine/src/test/java/io/camunda/zeebe/engine/processing/deployment/CreateDeploymentMultiplePartitionsTest.java
@@ -22,7 +22,6 @@ import io.camunda.zeebe.protocol.record.value.CommandDistributionRecordValue;
 import io.camunda.zeebe.protocol.record.value.DeploymentRecordValue;
 import io.camunda.zeebe.protocol.record.value.deployment.DecisionRecordValue;
 import io.camunda.zeebe.protocol.record.value.deployment.DecisionRequirementsMetadataValue;
-import io.camunda.zeebe.protocol.record.value.deployment.DeploymentResource;
 import io.camunda.zeebe.protocol.record.value.deployment.ProcessMetadataValue;
 import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.record.RecordingExporterTestWatcher;
@@ -119,15 +118,6 @@ public final class CreateDeploymentMultiplePartitionsTest {
         .getPartitionIds()
         .forEach(
             partitionId -> {
-              if (DEPLOYMENT_PARTITION == partitionId) {
-                assertDeploymentEventResources(
-                    partitionId,
-                    DeploymentIntent.CREATED,
-                    deployment.getKey(),
-                    (createdDeployment) -> assertDeploymentRecord(deployment, createdDeployment));
-                return;
-              }
-
               assertDeploymentEventResources(
                   partitionId,
                   DeploymentIntent.CREATED,
@@ -416,24 +406,6 @@ public final class CreateDeploymentMultiplePartitionsTest {
 
     assertThat(repeatedDrgs.size()).isEqualTo(PARTITION_COUNT - 1);
     repeatedDrgs.forEach(r -> assertDifferentDrg(originalDrg.get(0), r));
-  }
-
-  private void assertDeploymentRecord(
-      final Record<DeploymentRecordValue> deployment,
-      final Record<DeploymentRecordValue> createdDeployment) {
-    final DeploymentResource resource = createdDeployment.getValue().getResources().get(0);
-
-    Assertions.assertThat(resource).hasResource(bpmnXml(PROCESS));
-
-    final List<ProcessMetadataValue> deployedProcesses =
-        createdDeployment.getValue().getProcessesMetadata();
-
-    assertThat(deployedProcesses).hasSize(1);
-    Assertions.assertThat(deployedProcesses.get(0))
-        .hasBpmnProcessId("shouldCreateDeploymentOnAllPartitions")
-        .hasVersion(1)
-        .hasProcessDefinitionKey(getDeployedProcess(deployment, 0).getProcessDefinitionKey())
-        .hasResourceName("process.bpmn");
   }
 
   private void assertDeploymentRecordWithoutResources(
